+++ NEWS.md
@@ -1,5 +1,6 @@
 # Version 1.0.9000.9002
-* Renamed `refDoseBeta` slot to `ref_dose_beta` in `DualEndpointBeta` class.
+* Renamed `refDoseBeta` slot to `ref_dose_beta` in `DualEndpointEmax` class.
+* Renamed `refDoseEmax` slot to `ref_dose_emax` in `DualEndpointBeta` class.
 * Replaced `smooth` character vector argument of `DualEndpointRW()` to `rw1` flag. 
 * Added new `ModelParamsNormal` class to represent parameters of bivariate normal
   distribution.